Record data field in JSON RPC Error object
The JSON RPC Specification includes 3 fields in the Error object:
code, message, & data. Currently, the `rpc.jsonError` (in geth)
object holds all three fields, but when is turned into a string
via `Error`, only the message field is returned.

This causes difficulties when trying to get information from the
engine API. The engine API uses the `With` pattern which stores
the extra information in the data field.

This commit checks if the data field exists and if it does,
includes it in the error message.